split service module and controller module
