+++ service-module/src/main/java/com/luv2code/springboot/demo/models/repositories/OrderItemRepository.java
@@ -1,0 +1,9 @@
+package com.luv2code.springboot.demo.models.repositories;
+
+import com.luv2code.springboot.demo.models.entities.OrderItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+}
