+++ service-module/src/main/java/com/luv2code/springboot/demo/models/repositories/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.luv2code.springboot.demo.models.repositories;
+import com.luv2code.springboot.demo.models.entities.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Integer> {
+}
