+++ service-module/src/main/java/com/luv2code/springboot/demo/models/repositories/PurchaseOrderRepository.java
@@ -1,0 +1,9 @@
+package com.luv2code.springboot.demo.models.repositories;
+
+import com.luv2code.springboot.demo.models.entities.PurchaseOrder;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PurchaseOrderRepository extends JpaRepository<PurchaseOrder, Integer> {
+}
