+++ service-module/src/main/java/com/luv2code/springboot/demo/models/services/OrderItemService.java
@@ -1,0 +1,33 @@
+package com.luv2code.springboot.demo.models.services;
+
+
+import com.luv2code.springboot.demo.models.entities.OrderItem;
+import com.luv2code.springboot.demo.models.repositories.OrderItemRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class OrderItemService {
+
+    @Autowired
+    private OrderItemRepository orderItemRepository;
+
+    public List<OrderItem> findAll() {
+        return orderItemRepository.findAll();
+    }
+
+    public Optional<OrderItem> findById(Integer id) {
+        return orderItemRepository.findById(id);
+    }
+
+    public OrderItem save(OrderItem orderItem) {
+        return orderItemRepository.save(orderItem);
+    }
+
+    public void deleteById(Integer id) {
+        orderItemRepository.deleteById(id);
+    }
+}
