--- src/main/java/com/luv2code/springboot/demo/mycoolapp/controllers/ProductController.java
@@ -1,64 +1,0 @@
-package com.luv2code.springboot.demo.mycoolapp.controllers;
-
-import com.luv2code.springboot.demo.mycoolapp.models.entities.Product;
-import com.luv2code.springboot.demo.mycoolapp.models.services.ProductService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Optional;
-
-@RestController
-@RequestMapping("/api/v1/products")
-public class ProductController {
-
-    @Autowired
-    private ProductService productService;
-
-    @GetMapping
-    public List<Product> getAllProducts() {
-        return productService.findAll();
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<Product> getProductById(@PathVariable Integer id) {
-        Optional<Product> product = productService.findById(id);
-        if (product.isPresent()) {
-            return ResponseEntity.ok(product.get());
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @PostMapping
-    public Product createProduct(@RequestBody Product product) {
-        return productService.save(product);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Integer id, @RequestBody Product productDetails) {
-        Optional<Product> product = productService.findById(id);
-        if (product.isPresent()) {
-            Product updatedProduct = product.get();
-            updatedProduct.setName(productDetails.getName());
-            updatedProduct.setDescription(productDetails.getDescription());
-            updatedProduct.setPrice(productDetails.getPrice());
-            productService.save(updatedProduct);
-            return ResponseEntity.ok(updatedProduct);
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteProduct(@PathVariable Integer id) {
-        Optional<Product> product = productService.findById(id);
-        if (product.isPresent()) {
-            productService.deleteById(id);
-            return ResponseEntity.noContent().build();
-        } else {
-            return ResponseEntity.notFound().build();
-        }
-    }
-}
