--- src/main/java/com/luv2code/springboot/demo/mycoolapp/models/entities/Customer.java
@@ -1,104 +1,0 @@
-package com.luv2code.springboot.demo.mycoolapp.models.entities;
-
-import jakarta.persistence.*;
-
-import java.util.List;
-import java.util.Objects;
-
-
-@Entity
-public class Customer {
-
-    @Id
-    @SequenceGenerator(
-            name = "customer_id_sequence",
-            sequenceName = "customer_id_sequence",
-            allocationSize = 1
-    )
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "customer_id_sequence"
-    )
-
-    private Integer id;
-    private String name;
-    private String email;
-    private Integer age;
-
-    @OneToMany(mappedBy = "customer", cascade = CascadeType.ALL, fetch = FetchType.LAZY) // Ensure purchaseOrders are eagerly fetched
-    private List<PurchaseOrder> purchaseOrders;
-
-
-
-    public Customer(Integer id,
-                    String name,
-                    String email,
-                    Integer age) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-        this.age = age;
-    }
-
-    public Customer() {
-
-    }
-
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Customer customer = (Customer) o;
-        return Objects.equals(id, customer.id) && Objects.equals(name, customer.name) && Objects.equals(email, customer.email) && Objects.equals(age, customer.age);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, email, age);
-    }
-
-    @Override
-    public String toString() {
-        return "Customer{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", email='" + email + '\'' +
-                ", age='" + age + '\'' +
-                '}';
-    }
-}
-
-
-
